+++ src/commands/slashCommands/user/settings.js
@@ -1,10 +1,10 @@
 const { SlashCommandBuilder, InteractionContextType } = require('discord.js');
 const { linkSubCommand, linkSubFunction } = require('../../subcommands/link');
 const { messageSubCommand, messageSubFunction } = require('../../subcommands/message');
-const { embedColorSubComand, embedColorSubFunction } = require('../../subcommands/embedColor');
 const { deleteSubCommand, deleteSubFunction } = require('../../subcommands/delete');
 const { unlinkSubCommand, unlinkSubFunction } = require('../../subcommands/unlink');
 const { languageSubCommand, languageSubFunction } = require('../../subcommands/language');
+const { embedColorCustomSubCommand, embedColorCustomSubFunction, autocomplete: embedColorCustomAutocomplete } = require('../../subcommands/embedColor/embedColorCustom');
 
 module.exports = {
     premium: true,
@@ -15,18 +15,41 @@ module.exports = {
         .setDescriptionLocalizations({ nl: 'Kies een setting om hem te veranderen.' })
         .addSubcommand(linkSubCommand)
         .addSubcommand(messageSubCommand)
-        .addSubcommand(embedColorSubComand)
         .addSubcommand(deleteSubCommand)
         .addSubcommand(unlinkSubCommand)
-        .addSubcommand(languageSubCommand),
+        .addSubcommand(languageSubCommand)
+        .addSubcommandGroup(group => group
+            .setName('embedcolor')
+            .setDescription('Adjust embed color settings')
+            .addSubcommand(embedColorCustomSubCommand)
+        ),
 
     async execute(interaction, auth, userData) {
+
         let subCommand = interaction.options.getSubcommand();
-        if (subCommand === 'link') return linkSubFunction(interaction, auth, userData);
-        if (subCommand === 'message') return messageSubFunction(interaction, auth, userData);
-        if (subCommand === 'embedcolor') return embedColorSubFunction(interaction, auth, userData);
-        if (subCommand === 'delete') return deleteSubFunction(interaction, auth, userData);
-        if (subCommand === 'unlink') return unlinkSubFunction(interaction, auth, userData);
-        if (subCommand === 'language') return languageSubFunction(interaction, auth, userData);
+
+        const subCommandMap = {
+            link: (interaction, userData) => linkSubFunction(interaction, auth, userData),
+            message: messageSubFunction,
+            delete: deleteSubFunction,
+            unlink: unlinkSubFunction,
+            language: languageSubFunction,
+            custom: embedColorCustomSubFunction,
+        };
+
+        const subCommandFunction = subCommandMap[subCommand];
+        if (subCommandFunction) {
+            return subCommandFunction(interaction, userData);
+        } else {
+            console.error(`Unknown subcommand: ${subCommand}`);
+            return interaction.reply({ content: 'Unknown subcommand', ephemeral: true });
+        }
+    },
+
+    autocomplete(interaction) {
+        const subCommand = interaction.options.getSubcommand();
+        if (subCommand === 'custom') {
+            return embedColorCustomAutocomplete(interaction);
+        }
     }
 }